+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - refactor(Autocompleter): split abstract Autocompleter control vs Autocompleter form control
   - feat(Autocompleter): implement Autocompleter cell control
   - feat(IconButton): implement `disabled` API
+  - fix(Form Controls): ensure label use specified MDS fonts
 </details>
 
 ## 0.80.1