+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
+  - feat(Select): consumes `keydown` event when dropdown open
 </details>
 
 ## 0.81.0